requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
mouseinfo==0.1.3
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyautogui==0.9.54
pydantic==2.10.1
pydantic-core==2.27.1
pygetwindow==0.0.9
pygments==2.18.0
pymsgbox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-cocoa==10.3.1
pyobjc-framework-quartz==10.3.1
pyperclip==1.9.0
pyrect==0.2.0
pyscreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pyzmq==26.2.0
rubicon-objc==0.4.9
rumps==0.4.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
